Moving arith::nl::Inference to arith::Inference.
